Fix `dnf` removing kernel (YIKES!)
`dnf` dependency resolution attempts to remove `kernel` `kernel-core` `kernel-modules` and `kernel-modules-core`. This will brick the system.

This behavior is definitely not intended and indicates an issue with DNF's dependency resolution in this specific uninstallation context, or perhaps an overzealous dependency declaration from one of the CyberPanel components (like lsphp packages). DNF might perceive these core kernel components as "unused dependencies" after a primary package (like lsphp*) is slated for removal, which is a dangerous misinterpretation.